requirements.txt
Flask==1.1.4
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate==2.6.0
Flask-Script
Flask-Cors
requests
psycopg2
pika==1.1.0
python-dotenv
markupsafe==2.0.1